refactor docker intrinsic
