+++ src/rust/engine/src/intrinsics/mod.rs
@@ -31,11 +31,9 @@ use dep_inference::python::ParsedPythonDependencies;
 use futures::future::{BoxFuture, FutureExt, TryFutureExt};
 use futures::try_join;
 use indexmap::IndexMap;
-use pyo3::types::PyString;
 use pyo3::{IntoPy, PyAny, PyRef, Python, ToPyObject};
 use tokio::process;
 
-use docker::docker::{ImagePullPolicy, ImagePullScope, DOCKER, IMAGE_PULL_CACHE};
 use fs::{
     DigestTrie, DirectoryDigest, Entry, GlobMatching, PathStat, RelativePath, SymlinkBehavior,
     TypedPath,
@@ -44,7 +42,7 @@ use hashing::{Digest, EMPTY_DIGEST};
 use process_execution::local::{
     apply_chroot, create_sandbox, prepare_workdir, setup_run_sh_script, KeepSandboxes,
 };
-use process_execution::{ManagedChild, Platform, ProcessExecutionStrategy};
+use process_execution::{ManagedChild, ProcessExecutionStrategy};
 use rule_graph::{DependencyKey, RuleId};
 use stdio::TryCloneAsFile;
 use store::{SnapshotOps, Store, SubsetParams};
@@ -53,6 +51,10 @@ use crate::externs::dep_inference::PyNativeDependenciesRequest;
 use workunit_store::{in_workunit, Level};
 
 use grpc_util::prost::MessageExt;
+
+mod docker;
+
+use self::docker::docker_resolve_image;
 
 type IntrinsicFn =
     Box<dyn Fn(Context, Vec<Value>) -> BoxFuture<'static, NodeResult<Value>> + Send + Sync>;
@@ -791,60 +793,6 @@ fn interactive_process(
     )
     }
   ).boxed()
-}
-
-fn docker_resolve_image(
-    context: Context,
-    args: Vec<Value>,
-) -> BoxFuture<'static, NodeResult<Value>> {
-    async move {
-        let types = &context.core.types;
-        let docker_resolve_image_result = types.docker_resolve_image_result;
-
-        let (image_name, platform) = Python::with_gil(|py| {
-            let py_docker_request = (*args[0]).as_ref(py);
-            let image_name: String = externs::getattr(py_docker_request, "image_name").unwrap();
-            let platform: String = externs::getattr(py_docker_request, "platform").unwrap();
-            (image_name, platform)
-        });
-
-        let platform = Platform::try_from(platform)?;
-
-        let docker = DOCKER.get().await?;
-        let image_pull_scope = ImagePullScope::new(context.session.build_id());
-
-        // Ensure that the image has been pulled.
-        IMAGE_PULL_CACHE
-            .pull_image(
-                docker,
-                &context.core.executor,
-                &image_name,
-                &platform,
-                image_pull_scope,
-                ImagePullPolicy::OnlyIfLatestOrMissing,
-            )
-            .await
-            .map_err(|err| format!("Failed to pull image `{image_name}`: {err}"))?;
-
-        let image_metadata = docker.inspect_image(&image_name).await.map_err(|err| {
-            format!(
-                "Failed to resolve image ID for image `{}`: {:?}",
-                &image_name, err
-            )
-        })?;
-        let image_id = image_metadata
-            .id
-            .ok_or_else(|| format!("Image does not exist: `{}`", &image_name))?;
-
-        Ok(Python::with_gil(|py| {
-            externs::unsafe_call(
-                py,
-                docker_resolve_image_result,
-                &[Value::from(PyString::new(py, &image_id).to_object(py))],
-            )
-        }))
-    }
-    .boxed()
 }
 
 struct PreparedInferenceRequest {
